Better Printing

+++ src/cli/cli.py
@@ -11,6 +11,7 @@ from src.containers.container_manager_client import ContainerManagerClient
 from src.system.update import update, get_newest_supported_version
 from src.system.state import frozen
 from src.globals import VERSION
+from src.system.multithreading import SpinningTask
 
 CONTAINER_NAME_REGEX = r"""\w+"""
 
@@ -116,16 +117,20 @@ update-repo [URL]
         name = cmd[0]
         comp = re.compile(CONTAINER_NAME_REGEX)
         if not comp.match(name):
-            self.out_stream.write(f"'{name}' is not a valid container name\n")
+            self.out_stream.write(f"'{name}' is not a valid container name.\n")
             return
+        if not self.container_manager.started(name):
+            self.start([name])
         self.container_manager.sftp(name)
 
     def view_files(self, cmd: List[str]) -> None:
         name = cmd[0]
         comp = re.compile(CONTAINER_NAME_REGEX)
         if not comp.match(name):
-            self.out_stream.write(f"'{name}' is not a valid container name\n")
+            self.out_stream.write(f"'{name}' is not a valid container name.\n")
             return
+        if not self.container_manager.started(name):
+            self.start([name])
         try:
             self.container_manager.view_files(name)
         except (FileNotFoundError, PermissionError):
@@ -145,8 +150,10 @@ update-repo [URL]
         name = cmd[0]
         comp = re.compile(CONTAINER_NAME_REGEX)
         if not comp.match(name):
-            self.out_stream.write(f"'{name}' is not a valid container name\n")
+            self.out_stream.write(f"'{name}' is not a valid container name.\n")
             return
+        if not self.container_manager.started(name):
+            self.start([name])
         self.container_manager.run_shell(name)
 
     def start(self, cmd: List[str]) -> None:
@@ -158,10 +165,11 @@ update-repo [URL]
         name = cmd[0]
         comp = re.compile(CONTAINER_NAME_REGEX)
         if not comp.match(name):
-            self.out_stream.write(f"'{name}' is not a valid container name\n")
+            self.out_stream.write(f"'{name}' is not a valid container name.\n")
             return
-        self.container_manager.start(name)
-        self.container_manager.run_command(name, ["cat", "/etc/motd"])
+
+        t = SpinningTask(f"Starting {name}", self.container_manager.start, (name, ), self.out_stream)
+        t.exec()
 
     def stop(self, cmd: List[str]) -> None:
         """
@@ -172,9 +180,13 @@ update-repo [URL]
         name = cmd[0]
         comp = re.compile(CONTAINER_NAME_REGEX)
         if not comp.match(name):
-            self.out_stream.write(f"'{name}' is not a valid container name\n")
+            self.out_stream.write(f"'{name}' is not a valid container name.\n")
             return
-        self.container_manager.stop(name)
+        if not self.container_manager.started(name):
+            self.out_stream.write(f"{name} is not started.\n")
+        else:
+            self.container_manager.stop(name)
+            self.out_stream.write("Done.\n")
 
     def kill(self, cmd: List[str]) -> None:
         """
@@ -185,9 +197,13 @@ update-repo [URL]
         name = cmd[0]
         comp = re.compile(CONTAINER_NAME_REGEX)
         if not comp.match(name):
-            self.out_stream.write(f"'{name}' is not a valid container name\n")
+            self.out_stream.write(f"'{name}' is not a valid container name.\n")
             return
-        self.container_manager.kill(name)
+        if not self.container_manager.started(name):
+            self.out_stream.write(f"{name} is not started.\n")
+        else:
+            self.container_manager.kill(name)
+            self.out_stream.write("Done.\n")
 
     def run(self, cmd: List[str]) -> None:
         """
@@ -201,8 +217,10 @@ update-repo [URL]
         container_name, command = cmd[0], cmd[1:]
         comp = re.compile(CONTAINER_NAME_REGEX)
         if not comp.match(container_name):
-            self.out_stream.write(f"'{container_name}' is not a valid container name\n")
+            self.out_stream.write(f"'{container_name}' is not a valid container name.\n")
             return
+        if not self.container_manager.started(container_name):
+            self.start([container_name])
         self.container_manager.run_command(container_name, command)
 
     def send_file(self, cmd: List[str]) -> None:
@@ -221,10 +239,14 @@ update-repo [URL]
 
         comp = re.compile(CONTAINER_NAME_REGEX)
         if not comp.match(container_name):
-            self.out_stream.write(f"'{container_name}' is not a valid container name\n")
+            self.out_stream.write(f"'{container_name}' is not a valid container name.\n")
             return
+        if not self.container_manager.started(container_name):
+            self.start([container_name])
 
-        self.container_manager.put_file(container_name, local_file, remote_file)
+        p = f"Copying '{local_file}' -> '{remote_file if remote_file else '~'}'"
+        t = SpinningTask(p, self.container_manager.put_file, (container_name, local_file, remote_file))
+        t.exec()
 
     def get_file(self, cmd: List[str]) -> None:
         """
@@ -242,10 +264,14 @@ update-repo [URL]
 
         comp = re.compile(CONTAINER_NAME_REGEX)
         if not comp.match(container_name):
-            self.out_stream.write(f"'{container_name}' is not a valid container name\n")
+            self.out_stream.write(f"'{container_name}' is not a valid container name.\n")
             return
+        if not self.container_manager.started(container_name):
+            self.start([container_name])
 
-        self.container_manager.get_file(container_name, remote_file, local_file)
+        p = f"Copying '{remote_file}' -> '{local_file if local_file else '.'}'"
+        t = SpinningTask(p, self.container_manager.get_file, (container_name, remote_file, local_file))
+        t.exec()
 
     def install(self, cmd: List[str]) -> None:
         """
@@ -259,9 +285,10 @@ update-repo [URL]
         archive_path_str, container_name = cmd[0], cmd[1]
         comp = re.compile(CONTAINER_NAME_REGEX)
         if not comp.match(container_name):
-            self.out_stream.write(f"'{container_name}' is not a valid container name\n")
+            self.out_stream.write(f"'{container_name}' is not a valid container name.\n")
             return
-        self.container_manager.install(archive_path_str, container_name)
+        t = SpinningTask(f"Installing {container_name}", self.container_manager.install, (archive_path_str, container_name))
+        t.exec()
 
     def delete(self, cmd: List[str]) -> None:
         """
@@ -272,9 +299,13 @@ update-repo [URL]
         container_name = cmd[0]
         comp = re.compile(CONTAINER_NAME_REGEX)
         if not comp.match(container_name):
-            self.out_stream.write(f"'{container_name}' is not a valid container name\n")
+            self.out_stream.write(f"'{container_name}' is not a valid container name.\n")
             return
-        self.container_manager.delete(container_name)
+        if self.container_manager.started(container_name):
+            self.out_stream.write(f"Please stop {container_name} before trying to delete it.")
+        else:
+            self.out_stream.write("Command not yet supported")
+            # self.container_manager.delete(container_name)
 
     def download(self, cmd: List[str]) -> None:  # pylint: disable=unused-argument
         """
@@ -338,7 +369,8 @@ update-repo [URL]
 
         :param cmd: The rest of the command sent
         """
-        self.container_manager.ping()
+        t = self.container_manager.ping()
+        self.out_stream.write(f"Got OK in {t:.5f} seconds.\n")
 
     def ssh_address(self, cmd: List[str]) -> None:  # pylint: disable=unused-argument
         """
@@ -349,8 +381,10 @@ update-repo [URL]
         container_name = cmd[0]
         comp = re.compile(CONTAINER_NAME_REGEX)
         if not comp.match(container_name):
-            self.out_stream.write(f"'{container_name}' is not a valid container name\n")
+            self.out_stream.write(f"'{container_name}' is not a valid container name.\n")
             return
+        if not self.container_manager.started(container_name):
+            self.out_stream.write(f"{container_name} is not started.\n")
         self.out_stream.write(str(self.container_manager.ssh_address(container_name)))
 
     def update(self, cmd: List[str]) -> None:
